pyproject.toml
[project]
name = "autogen-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.5.1",
    "autogen-ext[ollama]>=0.5.1",
    "markdownify>=1.1.0",
    "openai>=1.71.0",
    "python-docx>=1.1.2",
]
